Don't have test_mkalias_relative run if sys.prefix does not exist.  Backport of fix for bug #1077302.

+++ Lib/test/test_macostools.py
@@ -71,6 +71,8 @@ class TestMacostools(unittest.TestCase):
         self.assertEqual(fss.as_pathname(), os.path.realpath(test_support.TESTFN))
 
     def test_mkalias_relative(self):
+        if not os.path.exists(sys.prefix):
+            return
         try:
             os.unlink(TESTFN2)
         except:
